// frontend/src/client/types.gen.ts
export type UserPublic = {
    id: string;
    email: string;
    is_active: boolean;
    is_superuser: boolean;
    is_verified: boolean;
    full_name?: (string | null);
};